backend/requirements.txt
django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
spacy==3.6.0
transformers==4.30.2
torch==2.0.1
pdf2image==1.16.3
pytesseract==0.3.10
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
PyPDF2==3.0.1
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.23.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
pillow==9.5.0
huggingface_hub==0.16.4 